MRViewer/MRRibbonRegisterItem.h
