+++ source/MRViewer/MRRibbonRegisterItem.cpp
@@ -1,0 +1,17 @@
+#include "MRRibbonRegisterItem.h"
+#include "MRRibbonSchema.h"
+
+namespace MR
+{
+
+RibbonMenuItemAdder::RibbonMenuItemAdder( std::shared_ptr<RibbonMenuItem> item ) : item_( std::move( item ) )
+{
+    RibbonSchemaHolder::addItem( item_ );
+}
+
+RibbonMenuItemAdder::~RibbonMenuItemAdder()
+{
+    RibbonSchemaHolder::delItem( item_ );
+}
+
+} //namespace MR
