+++ source/MRViewer/MRViewer.vcxproj
@@ -44,6 +44,7 @@
     <ClCompile Include="MRRenderDefaultObjects.cpp" />
     <ClCompile Include="MRRibbonLayoutConfig.cpp" />
     <ClCompile Include="MRRibbonNotification.cpp" />
+    <ClCompile Include="MRRibbonRegisterItem.cpp" />
     <ClCompile Include="MRRibbonSceneObjectsListDrawer.cpp" />
     <ClCompile Include="MRUISaveChangesPopup.cpp" />
     <ClCompile Include="MRSceneCache.cpp" />
@@ -181,6 +182,7 @@
     <ClInclude Include="MRProjectMeshAttributes.h" />
     <ClInclude Include="MRRibbonLayoutConfig.h" />
     <ClInclude Include="MRRibbonNotification.h" />
+    <ClInclude Include="MRRibbonRegisterItem.h" />
     <ClInclude Include="MRRibbonSceneObjectsListDrawer.h" />
     <ClInclude Include="MRMruFormatParameters.h" />
     <ClInclude Include="MRUISaveChangesPopup.h" />
